+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Exception ``NoSolverSelected`` raised when the solver kwarg is missing
 - Starter set of QP solvers as optional dependencies
+- Test exceptions raised by `solve_ls` and `solve_qp`
 
 ### Changed